+++ projects/ngx-scrollbar/mat/src/public_api.ts
@@ -1,0 +1,1 @@
+export * from './mat-select-viewport';